cleanup add_argument functions
