+++ precimed/common/utils_cli.py
@@ -4,6 +4,60 @@ import collections
 import argparse
 import common.libbgmg
 
+GO_EXTEND_BP_DEFAULT = 10000
+
+def parser_add_argument_bim_file(parser):
+    parser.add_argument("--bim-file", type=str, default=None, help="plink bim file (required argument); "
+        "defines the reference set of SNPs used for the analysis. "
+        "Marker names must not have duplicated entries. "
+        "May contain symbol '@', which will be replaced by an actual chromosome label. ")
+
+def parser_add_argument_ld_file(parser):
+    parser.add_argument("--ld-file", type=str, default=None, help="file with linkage disequilibrium information, "
+        "generated via 'mixer.py ld' command (required argument); "
+        "may contain symbol '@', similarly to --bim-file argument. ")
+
+def parser_add_argument_chr2use(parser):
+    parser.add_argument("--chr2use", type=str, default="1-22", help="chromosome labels to use (default: %(default)s); "
+        "chromosome must be labeled by integer, i.e. X and Y are not acceptable; example of valid arguments: '1,2,3' or '1-4,12,16-20'")
+
+def parser_add_argument_extract_and_exclude(parser):
+    parser.add_argument('--extract', type=str, default="", help="(optional) File with variants to include in the analysis. By default, all variants are included. This applies to GWAS tag SNPs, however LD is still computed towards the entire reference provided by --bim-file. This applies before --exclude, so a variant listed in --exclude won't be re-introduced if it's also present in --extract list.")
+    parser.add_argument('--exclude', type=str, default="", help="(optional) File with variants to exclude from the analysis.")
+    parser.add_argument('--exclude-ranges', type=str, default=['MHC'], nargs='+',
+        help='(default: %(default)s) exclude SNPs in ranges of base pair position; '
+        'the syntax is chr:from-to, for example 6:25000000-35000000; '
+        'multiple regions can be excluded; '
+        '"chr" prefix prior to chromosome label, as well as KB and MB suffices are allowed, e.g. chr6:25-35MB is a valid exclusion range. '
+        'Some special case regions are also supported, for example "--exclude-ranges MHC APOE".'
+        'To overwrite the default, pass "--exclude-ranges []".')
+
+def parser_add_argument_use_complete_tag_indices(parser):
+    parser.add_argument('--use-complete-tag-indices', default=False, action="store_true", help=argparse.SUPPRESS) #"advanced option (expert use only); MiXeR software has a notion of 'tag SNPs', which are the subset of the SNPs in --bim-file with well defined GWAS z-score(s); this saves memory used by LD r2 storage, which became a rectangular matrix (LD r2 between two SNPs from the --bim-file is stored only if one of the SNPs is a tag SNP). Setting this flag enforces all --bim-file SNPs to be treated as tag SNPs, some will have undefined GWAS z-score(s), and the complete LD matrix will be save in memory.")
+
+def parser_add_argument_allow_albiguous_snps(parser):
+    parser.add_argument('--allow-ambiguous-snps', default=False, action="store_true", help="advanced option (expert use only); a flag allowing to include A/T and C/G SNPs in fit procedure.")
+
+def parser_add_argument_savelib_and_loadlib(parser):
+    parser.add_argument("--savelib-file", type=str, default=None, help="optional path to save the state of the libbgmg object (for faster initialization with --loadlib-file). It is highly recommended to use --use-complete-tag-indices together with --savelib-file.")
+    parser.add_argument("--loadlib-file", type=str, default=None, help="optional path to load libbgmg object; "
+        "there are several limitations if you use this option: "
+        "--bim-file argument must be the same as what was used with --savelib-file; "
+        "for plsa analysis --chr2use can be different as each chromosome has its own .bin file; "
+        "--ld-file and --use-complete-tag-indices arguments will be ignored.")
+
+def parser_add_argument_annot_file(parser, annot_file_test=False):
+    parser.add_argument("--annot-file", type=str, default=None, help="(optional) path to binary annotations in LD score regression format, i.e. <path>/baseline.@.annot.gz for fitting enrichment model model. This must include the first column with all ones ('base' annotation category covering the entire genome).")
+    if annot_file_test:
+        parser.add_argument("--annot-file-test", type=str, default=None, help="(optional) path to binary annotations in LD score regression format, i.e. <path>/baseline.@.annot.gz for evaluating enrichment.  If provided, enrichment scores computed for --annot-file-test will be saved to <out>.annot_test_enrich.csv.")
+
+def parser_add_argument_go_file(parser, go_file_test=False):
+    # --go-file and --go-file-test must contain columns 'GO', 'GENE', 'CHR', 'FROM', 'TO'
+    parser.add_argument("--go-file", type=str, default=None, help="(optional) path to GO antology file for fitting enrichment model model. The format is described in the documentation. 'base' category that covers entire genome will be added automatically.")
+    parser.add_argument("--go-extend-bp", type=int, default=GO_EXTEND_BP_DEFAULT, help="extends each gene by this many base pairs, defining a symmetric window up and downstream (default: %(default)s)")
+    if go_file_test:
+        parser.add_argument("--go-file-test", type=str, default=None, help="(optional) path to an additional  GO antology file for evaluating enrichment, same convention as --go-file. If provided, enrichment scores computed for --go-test-file will be saved to <out>.go_test_enrich.csv")
+
 def check_input_file(args, argname, chri=None):
     arg_dict = vars(args)
     if (argname in arg_dict) and arg_dict[argname]:
